// ProjectSelene.Modloader/src/moddb/generated/selene-api.ts
export interface ModDto {
    id: string;
    name: string;
    description: string;
    author: string;
    version: string;
    versions: string[];
}